move sort control components to library
