+++ src/renderer/screens/OpenGov/Tracks/index.tsx
@@ -18,7 +18,7 @@ import { ButtonPrimaryInvert } from '@/renderer/kits/Buttons/ButtonPrimaryInvert
 import { TrackRow } from './TrackRow';
 import { Scrollable, StatsFooter } from '@/renderer/library/styles';
 import { renderPlaceholders } from '@/renderer/library/utils';
-import { ControlsWrapper, SortControlButton } from '@/renderer/utils/common';
+import { ControlsWrapper, SortControlButton } from '@app/library/SortControls';
 import type { HelpItemKey } from '@/renderer/contexts/common/Help/types';
 import type { TracksProps } from '../types';
