+++ src/renderer/screens/Settings/Workspaces/WorkspaceRow.tsx
@@ -4,7 +4,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { WorkspaceRowWrapper } from '../Wrappers';
 import { faHashtag, faTrash, faLink } from '@fortawesome/free-solid-svg-icons';
-import { ControlsWrapper, SortControlButton } from '@/renderer/utils/common';
+import { ControlsWrapper, SortControlButton } from '@app/library/SortControls';
 import { useTooltip } from '@/renderer/contexts/common/Tooltip';
 import { useWorkspaces } from '@/renderer/contexts/settings/Workspaces';
 import { useOverlay } from '@/renderer/contexts/common/Overlay';
